requirements.txt
certifi==2021.5.30
Click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
opencv-python==4.5.3.56
Pillow==8.3.1
pytesseract==0.3.8
Werkzeug==2.0.1